AssessmentServiceTests working.
