+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppstier/service/AssessmentApiServiceTest.kt
@@ -1,14 +1,23 @@
 package uk.gov.justice.digital.hmpps.hmppstier.service
 
 import io.mockk.clearMocks
+import io.mockk.every
 import io.mockk.junit5.MockKExtension
 import io.mockk.mockk
+import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
+import org.junit.jupiter.api.Nested
+import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
+import uk.gov.justice.digital.hmpps.hmppstier.client.Answer
 import uk.gov.justice.digital.hmpps.hmppstier.client.AssessmentApiClient
+import uk.gov.justice.digital.hmpps.hmppstier.client.OffenderAssessment
+import uk.gov.justice.digital.hmpps.hmppstier.client.Question
+import uk.gov.justice.digital.hmpps.hmppstier.domain.enums.AssessmentComplexityFactor
 import java.time.Clock
 import java.time.LocalDate
+import java.time.LocalDateTime
 import java.time.ZoneId
 
 @ExtendWith(MockKExtension::class)
@@ -23,7 +32,7 @@ internal class AssessmentApiServiceTest {
     clearMocks(assessmentApiClient)
   }
 
-  /*@Nested
+  @Nested
   @DisplayName("Get Complexity Answer Tests")
   inner class GetComplexityAnswerTests {
 
@@ -45,7 +54,7 @@ internal class AssessmentApiServiceTest {
           assessment.assessmentId
         )
       } returns complexityAnswers
-      val returnValue = assessmentService.getAssessmentComplexityAnswers(crn)
+      val returnValue = assessmentService.getAssessmentAnswers(assessment.assessmentId)
 
       assertThat(returnValue).hasSize(1)
       assertThat(returnValue).containsKey(AssessmentComplexityFactor.PARENTING_RESPONSIBILITIES)
@@ -69,7 +78,7 @@ internal class AssessmentApiServiceTest {
           assessment.assessmentId
         )
       } returns complexityAnswers
-      val returnValue = assessmentService.getAssessmentComplexityAnswers(crn)
+      val returnValue = assessmentService.getAssessmentAnswers(assessment.assessmentId)
 
       assertThat(returnValue).hasSize(1)
       assertThat(returnValue).containsKey(AssessmentComplexityFactor.PARENTING_RESPONSIBILITIES)
@@ -93,7 +102,7 @@ internal class AssessmentApiServiceTest {
           assessment.assessmentId
         )
       } returns complexityAnswers
-      val returnValue = assessmentService.getAssessmentComplexityAnswers(crn)
+      val returnValue = assessmentService.getAssessmentAnswers(assessment.assessmentId)
 
       assertThat(returnValue).hasSize(1)
       assertThat(returnValue).containsKey(AssessmentComplexityFactor.PARENTING_RESPONSIBILITIES)
@@ -117,7 +126,7 @@ internal class AssessmentApiServiceTest {
           assessment.assessmentId
         )
       } returns complexityAnswers
-      val returnValue = assessmentService.getAssessmentComplexityAnswers(crn)
+      val returnValue = assessmentService.getAssessmentAnswers(assessment.assessmentId)
 
       assertThat(returnValue).hasSize(1)
       assertThat(returnValue).containsKey(AssessmentComplexityFactor.PARENTING_RESPONSIBILITIES)
@@ -141,7 +150,7 @@ internal class AssessmentApiServiceTest {
           assessment.assessmentId
         )
       } returns complexityAnswers
-      val returnValue = assessmentService.getAssessmentComplexityAnswers(crn)
+      val returnValue = assessmentService.getAssessmentAnswers(assessment.assessmentId)
 
       assertThat(returnValue).isEmpty()
     }
@@ -164,7 +173,7 @@ internal class AssessmentApiServiceTest {
           assessment.assessmentId
         )
       } returns complexityAnswers
-      val returnValue = assessmentService.getAssessmentComplexityAnswers(crn)
+      val returnValue = assessmentService.getAssessmentAnswers(assessment.assessmentId)
 
       assertThat(returnValue).hasSize(1)
       assertThat(returnValue).containsKey(AssessmentComplexityFactor.PARENTING_RESPONSIBILITIES)
@@ -192,7 +201,7 @@ internal class AssessmentApiServiceTest {
           assessment.assessmentId
         )
       } returns complexityAnswers
-      val returnValue = assessmentService.getAssessmentComplexityAnswers(crn)
+      val returnValue = assessmentService.getAssessmentAnswers(assessment.assessmentId)
 
       assertThat(returnValue).hasSize(2)
       assertThat(returnValue).containsKey(AssessmentComplexityFactor.PARENTING_RESPONSIBILITIES)
@@ -212,98 +221,9 @@ internal class AssessmentApiServiceTest {
           assessment.assessmentId
         )
       } returns complexityAnswers
-      val returnValue = assessmentService.getAssessmentComplexityAnswers(crn)
+      val returnValue = assessmentService.getAssessmentAnswers(assessment.assessmentId)
 
       assertThat(returnValue).isEmpty()
     }
   }
-
-  @Nested
-  @DisplayName("Get Needs Tests")
-  inner class GetNeedsTests {
-
-    @Test
-    fun `Should return empty Map if no Needs`() {
-      val crn = "123"
-      val assessment = OffenderAssessment("1234", LocalDateTime.now(clock), null)
-      val needs = listOf<AssessmentNeed>()
-
-      every { assessmentApiClient.getAssessmentSummaries(crn) } returns listOf(assessment)
-      every { assessmentApiClient.getAssessmentNeeds(assessment.assessmentId) } returns needs
-      val returnValue = assessmentService.getAssessmentNeeds(crn)
-
-      assertThat(returnValue).isEmpty()
-    }
-
-    @Test
-    fun `Should return Needs`() {
-      val crn = "123"
-      val assessment = OffenderAssessment("1234", LocalDateTime.now(clock), null)
-      val needs = listOf(
-        AssessmentNeed(
-          Need.ACCOMMODATION,
-          NeedSeverity.NO_NEED
-        )
-      )
-
-      every { assessmentApiClient.getAssessmentSummaries(crn) } returns listOf(assessment)
-      every { assessmentApiClient.getAssessmentNeeds(assessment.assessmentId) } returns needs
-      val returnValue = assessmentService.getAssessmentNeeds(crn)
-
-      assertThat(returnValue).hasSize(1)
-      assertThat(returnValue).containsEntry(Need.ACCOMMODATION, NeedSeverity.NO_NEED)
-    }
-  }
-
-  @Nested
-  @DisplayName("Get recent Assessment Tests")
-  inner class GetRecentAssessmentTests {
-
-    @Test
-    fun `Should return true if inside Threshold`() {
-      val crn = "123"
-      val assessment = OffenderAssessment("1234", LocalDateTime.now(clock).minusWeeks(55), null)
-
-      every { assessmentApiClient.getAssessmentSummaries(crn) } returns listOf(assessment)
-      val returnValue = assessmentService.isAssessmentRecent(crn)
-
-      assertThat(returnValue).isTrue
-    }
-
-    @Test
-    fun `Should return false if outside Threshold`() {
-      val crn = "123"
-      val assessment = OffenderAssessment("1234", LocalDateTime.now(clock).minusWeeks(55).minusDays(1), null)
-      // more recent, but voided
-      val voidedAssessment = OffenderAssessment("1234", LocalDateTime.now(clock).minusWeeks(40), LocalDateTime.now(clock))
-
-      every { assessmentApiClient.getAssessmentSummaries(crn) } returns listOf(assessment, voidedAssessment)
-      val returnValue = assessmentService.isAssessmentRecent(crn)
-
-      assertThat(returnValue).isFalse
-    }
-
-    @Test
-    fun `Should throw if none valid`() {
-      val crn = "123"
-
-      every { assessmentApiClient.getAssessmentSummaries(crn) } returns listOf()
-
-      assertThrows(EntityNotFoundException::class.java) {
-        assessmentService.isAssessmentRecent(crn)
-      }
-    }
-
-    @Test
-    fun `Should throw if none valid with entries`() {
-      val crn = "123"
-      val assessment = OffenderAssessment("1234", LocalDateTime.now(clock).minusWeeks(55).minusDays(1), LocalDateTime.now(clock))
-
-      every { assessmentApiClient.getAssessmentSummaries(crn) } returns listOf(assessment)
-
-      assertThrows(EntityNotFoundException::class.java) {
-        assessmentService.isAssessmentRecent(crn)
-      }
-    }
-  }*/
 }
